+++ addons/twitch_irc_client/message_queue.gd
@@ -13,11 +13,6 @@ func _on_dispatch_timeout() -> void:
 	if not _outgoing_messages.is_empty():
 		var message: String = _outgoing_messages.pop_front()
 		dispatch_requested.emit(message + CRLF)
-		if owner.enable_log and OS.is_debug_build():
-			if message.begins_with("PASS"):
-				print("< PASS oauth:******************************")
-			else:
-				prints("<", message.strip_edges())
 
 
 func add(message: String) -> void: