+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+- BREAKING CHANGE: Remove `enable_log` property
+- Add `logger` signal
 - Add `TwitchIRCClient#is_connection_open` method
 
 ## 2023-02-01 - v0.1.1